requirements.txt
numpy>=1.20.2
matplotlib>=3.3.4
visualdl>=2.1.1
paddlepaddle==2.0.1